+++ client/src/components/Team.js
@@ -1,40 +1,5 @@
 import React from "react";
-
-function Website(props) {
-  let responses = [];
-  if (props.website) {
-    console.log(props.website);
-    responses.push(
-      <p>
-        <a href={props.website}>Website - Jtsai972.space</a>
-      </p>
-    );
-  }
-
-  if (props.github) {
-    console.log(props.website);
-    responses.push(
-      <p>
-        <a href={"https://github.com/" + props.github}>
-          Github - {props.github}
-        </a>
-      </p>
-    );
-  }
-
-  if (props.linkedin) {
-    console.log(props.website);
-    responses.push(
-      <p>
-        <a href={"https://www.linkedin.com/in/" + props.linkedin}>
-          LinkedIn - {props.linkedin}
-        </a>
-      </p>
-    );
-  }
-
-  return responses;
-}
+import Websites from "./About/Websites"
 
 function Team(props) {
   console.log(props);
@@ -45,7 +10,7 @@ function Team(props) {
         <img className="card-img-top" src={props.src} alt={props.name} />
         <div className="card-body">
           <h4 className="h5 card-title">{props.name}</h4>
-          <Website
+          <Websites
             website={props.website}
             github={props.github}
             linkedin={props.linkedin}
